Create Flow definition file for Lexical core
