+++ jest.config.js
@@ -86,8 +86,6 @@ module.exports = {
           '<rootDir>/packages/lexical/src/nodes/extended/LexicalQuoteNode.js',
         '^shared/canUseDOM$': '<rootDir>/packages/shared/src/canUseDOM.js',
         '^shared/environment$': '<rootDir>/packages/shared/src/environment.js',
-        '^shared/getPossibleDecoratorNode$':
-          '<rootDir>/packages/shared/src/getPossibleDecoratorNode.js',
         '^shared/invariant$': '<rootDir>/packages/shared/src/invariant.js',
         '^shared/useLayoutEffect$':
           '<rootDir>/packages/shared/src/useLayoutEffect.js',
