--- packages/shared/src/getPossibleDecoratorNode.js
@@ -1,60 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @flow strict
- */
-
-import type {ElementNode, LexicalNode, Point} from 'lexical';
-
-function resolveElement(
-  element: ElementNode,
-  isBackward: boolean,
-  focusOffset: number,
-): LexicalNode | null {
-  const parent = element.getParent();
-  let offset = focusOffset;
-  let block = element;
-  if (parent !== null) {
-    if (isBackward && focusOffset === 0) {
-      offset = block.getIndexWithinParent();
-      block = parent;
-    } else if (!isBackward && focusOffset === block.getChildrenSize()) {
-      offset = block.getIndexWithinParent() + 1;
-      block = parent;
-    }
-  }
-  return block.getChildAtIndex(isBackward ? offset - 1 : offset);
-}
-
-export default function getPossibleDecoratorNode(
-  focus: Point,
-  isBackward: boolean,
-): null | LexicalNode {
-  const focusOffset = focus.offset;
-  if (focus.type === 'element') {
-    const block = focus.getNode();
-    return resolveElement(block, isBackward, focusOffset);
-  } else {
-    const focusNode = focus.getNode();
-    if (
-      (isBackward && focusOffset === 0) ||
-      (!isBackward && focusOffset === focusNode.getTextContentSize())
-    ) {
-      const possibleNode = isBackward
-        ? focusNode.getPreviousSibling()
-        : focusNode.getNextSibling();
-      if (possibleNode === null) {
-        return resolveElement(
-          focusNode.getParentOrThrow(),
-          isBackward,
-          focusNode.getIndexWithinParent() + (isBackward ? 0 : 1),
-        );
-      }
-      return possibleNode;
-    }
-  }
-  return null;
-}
